backend/requirements.txt
alembic
backoff
cryptography
chromadb
fastapi[standard]
pinecone-client
psycopg2
psycopg2-binary
pyjwt
requests
sqlalchemy[asyncio]
temporalio
tweepy
uvicorn[standard]